+++ core/events/tx_error.go
@@ -102,6 +102,10 @@ func NewTxErrEvent(ctx context.Context, err error, partyID string, tx interface{
 		evt.evt.Transaction = &eventspb.TxErrorEvent_IssueSignatures{
 			IssueSignatures: tv,
 		}
+	case *commandspb.BatchMarketInstructions:
+		evt.evt.Transaction = &eventspb.TxErrorEvent_BatchMarketInstructions{
+			BatchMarketInstructions: tv,
+		}
 	case error: // unsupported command error
 		evt.evt.ErrMsg = fmt.Sprintf("%v - %v", err, tv)
 	}