+++ core/execution/engine.go
@@ -500,7 +500,8 @@ func (e *Engine) SubmitOrder(
 	ctx context.Context,
 	submission *types.OrderSubmission,
 	party string,
-	deterministicID string,
+	idgen IDGenerator,
+	orderID string,
 ) (confirmation *types.OrderConfirmation, returnedErr error) {
 	timer := metrics.NewTimeCounter(submission.MarketID, "execution", "SubmitOrder")
 	defer func() {
@@ -517,7 +518,8 @@ func (e *Engine) SubmitOrder(
 	}
 
 	metrics.OrderGaugeAdd(1, submission.MarketID)
-	conf, err := mkt.SubmitOrder(ctx, submission, party, deterministicID)
+	conf, err := mkt.SubmitOrderWithIDGeneratorAndOrderID(
+		ctx, submission, party, idgen, orderID)
 	if err != nil {
 		return nil, err
 	}
@@ -529,8 +531,11 @@ func (e *Engine) SubmitOrder(
 
 // AmendOrder takes order amendment details and attempts to amend the order
 // if it exists and is in a editable state.
-func (e *Engine) AmendOrder(ctx context.Context, amendment *types.OrderAmendment, party string,
-	deterministicID string,
+func (e *Engine) AmendOrder(
+	ctx context.Context,
+	amendment *types.OrderAmendment,
+	party string,
+	idgen IDGenerator,
 ) (confirmation *types.OrderConfirmation, returnedErr error) {
 	timer := metrics.NewTimeCounter(amendment.MarketID, "execution", "AmendOrder")
 	defer func() {
@@ -546,7 +551,7 @@ func (e *Engine) AmendOrder(ctx context.Context, amendment *types.OrderAmendment
 		return nil, types.ErrInvalidMarketID
 	}
 
-	conf, err := mkt.AmendOrder(ctx, amendment, party, deterministicID)
+	conf, err := mkt.AmendOrderWithIDGenerator(ctx, amendment, party, idgen)
 	if err != nil {
 		return nil, err
 	}
@@ -577,7 +582,12 @@ func (e *Engine) decrementOrderGaugeMetrics(
 }
 
 // CancelOrder takes order details and attempts to cancel if it exists in matching engine, stores etc.
-func (e *Engine) CancelOrder(ctx context.Context, cancel *types.OrderCancellation, party string, deterministicID string) (_ []*types.OrderCancellationConfirmation, returnedErr error) {
+func (e *Engine) CancelOrder(
+	ctx context.Context,
+	cancel *types.OrderCancellation,
+	party string,
+	idgen IDGenerator,
+) (_ []*types.OrderCancellationConfirmation, returnedErr error) {
 	timer := metrics.NewTimeCounter(cancel.MarketID, "execution", "CancelOrder")
 	defer func() {
 		timer.EngineTimeCounterAdd()
@@ -594,19 +604,23 @@ func (e *Engine) CancelOrder(ctx context.Context, cancel *types.OrderCancellatio
 
 	if len(cancel.MarketID) > 0 {
 		if len(cancel.OrderID) > 0 {
-			return e.cancelOrder(ctx, party, cancel.MarketID, cancel.OrderID, deterministicID)
+			return e.cancelOrder(ctx, party, cancel.MarketID, cancel.OrderID, idgen)
 		}
 		return e.cancelOrderByMarket(ctx, party, cancel.MarketID)
 	}
 	return e.cancelAllPartyOrders(ctx, party)
 }
 
-func (e *Engine) cancelOrder(ctx context.Context, party, market, orderID string, deterministicID string) ([]*types.OrderCancellationConfirmation, error) {
+func (e *Engine) cancelOrder(
+	ctx context.Context,
+	party, market, orderID string,
+	idgen IDGenerator,
+) ([]*types.OrderCancellationConfirmation, error) {
 	mkt, ok := e.markets[market]
 	if !ok {
 		return nil, types.ErrInvalidMarketID
 	}
-	conf, err := mkt.CancelOrder(ctx, party, orderID, deterministicID)
+	conf, err := mkt.CancelOrderWithIDGenerator(ctx, party, orderID, idgen)
 	if err != nil {
 		return nil, err
 	}
@@ -730,16 +744,11 @@ func (e *Engine) CancelLiquidityProvision(ctx context.Context, cancel *types.Liq
 func (e *Engine) OnTick(ctx context.Context, t time.Time) {
 	timer := metrics.NewTimeCounter("-", "execution", "OnTick")
 
-	evts := make([]events.Event, 0, len(e.marketsCpy))
-	for _, v := range e.marketsCpy {
-		evts = append(evts, events.NewMarketDataEvent(ctx, v.GetMarketData()))
-	}
-	e.broker.SendBatch(evts)
-
 	e.log.Debug("updating engine on new time update")
 
 	// notify markets of the time expiration
 	toDelete := []string{}
+	evts := make([]events.Event, 0, len(e.marketsCpy))
 	for _, mkt := range e.marketsCpy {
 		mkt := mkt
 		closing := mkt.OnTick(ctx, t)
@@ -748,7 +757,9 @@ func (e *Engine) OnTick(ctx context.Context, t time.Time) {
 				logging.MarketID(mkt.GetID()))
 			toDelete = append(toDelete, mkt.GetID())
 		}
+		evts = append(evts, events.NewMarketDataEvent(ctx, mkt.GetMarketData()))
 	}
+	e.broker.SendBatch(evts)
 
 	for _, id := range toDelete {
 		e.removeMarket(id)