+++ protos/vega/governance.pb.go
@@ -208,6 +208,7 @@ func (ProposalError) EnumDescriptor() ([]byte, []int) {
 //   - Passed -> Enacted.
 //   - Passed -> Failed.
 //   - Declined
+//
 // Rejected
 // Proposal can enter Failed state from any other state
 type Proposal_State int32
@@ -442,6 +443,7 @@ type InstrumentConfiguration struct {
 	// Product specification
 	//
 	// Types that are assignable to Product:
+	//
 	//	*InstrumentConfiguration_Future
 	Product isInstrumentConfiguration_Product `protobuf_oneof:"product"`
 }
@@ -536,6 +538,7 @@ type NewMarketConfiguration struct {
 	// New market risk model parameters
 	//
 	// Types that are assignable to RiskParameters:
+	//
 	//	*NewMarketConfiguration_Simple
 	//	*NewMarketConfiguration_LogNormal
 	RiskParameters isNewMarketConfiguration_RiskParameters `protobuf_oneof:"risk_parameters"`
@@ -780,6 +783,7 @@ type UpdateMarketConfiguration struct {
 	// Updated market risk model parameters
 	//
 	// Types that are assignable to RiskParameters:
+	//
 	//	*UpdateMarketConfiguration_Simple
 	//	*UpdateMarketConfiguration_LogNormal
 	RiskParameters isUpdateMarketConfiguration_RiskParameters `protobuf_oneof:"risk_parameters"`
@@ -895,6 +899,7 @@ type UpdateInstrumentConfiguration struct {
 	// Product specification
 	//
 	// Types that are assignable to Product:
+	//
 	//	*UpdateInstrumentConfiguration_Future
 	Product isUpdateInstrumentConfiguration_Product `protobuf_oneof:"product"`
 }
@@ -1262,6 +1267,7 @@ type ProposalTerms struct {
 	// Changes being proposed
 	//
 	// Types that are assignable to Change:
+	//
 	//	*ProposalTerms_UpdateMarket
 	//	*ProposalTerms_NewMarket
 	//	*ProposalTerms_UpdateNetworkParameter